Precise start location

+++ global_planner/src/astar.cpp
@@ -43,7 +43,7 @@ AStarExpansion::AStarExpansion(PotentialCalculator* p_calc, int xs, int ys) :
         Expander(p_calc, xs, ys) {
 }
 
-bool AStarExpansion::calculatePotentials(unsigned char* costs, int start_x, int start_y, int end_x, int end_y,
+bool AStarExpansion::calculatePotentials(unsigned char* costs, double start_x, double start_y, double end_x, double end_y,
                                         int cycles, float* potential) {
     queue_.clear();
     int start_i = toIndex(start_x, start_y);
